requirements.txt
transformers==4.40.1
accelerate==0.29.3
peft==0.10.0
trl==0.8.6
datasets
python-dotenv
torch